wip: - Moved celeb and movie test to controllers
